Class Client_context.simple_printer

A simple printer can be used to implement a printer as it is done in class Client_context_unix.unix_logger.

method error : ('a, 'b) lwt_format -> 'a

method warning : ('a, unit) lwt_format -> 'a

method message : ('a, unit) lwt_format -> 'a

method answer : ('a, unit) lwt_format -> 'a

method log : string -> ('a, unit) lwt_format -> 'a