Class type Client_context.io_wallet

Primitives for input, output and wallet. The general organisation of the code in this module is to provide small classes that are also combined into bigger classes. It allows different client library functions to only depend on some features, but not all, so that these functions can be used in places that only have access to these features.

inherit printer

inherit prompter

inherit wallet