Class type Client_context.prompter

prompter is a class of objects that provide input functions to request data from the end-user, whether normal inputs or passwords.

method prompt : ('a, string Tezos_base__TzPervasives.tzresult) lwt_format -> 'a

method prompt_password : ('a, Stdlib.Bytes.t Tezos_base__TzPervasives.tzresult) lwt_format -> 'a

method multiple_password_retries : bool

when multiple_password_retries is true, password prompt should retries more than once. true is the default value.