Module type Client_keys.SIGNER

val scheme : string

scheme is the name of the scheme implemented by this signer module.

val title : string

title is a one-line human readable description of the signer.

val description : string

description is a multi-line human readable description of the signer, that should include the format of key specifications.

val neuterize : sk_uri -> (pk_uri, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

neuterize sk is the corresponding pk.

val import_secret_key : io:Client_context.io_wallet -> pk_uri -> (Tezos_crypto.Signature.Public_key_hash.t * Tezos_crypto.Signature.Public_key.t option, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

import_secret_key ~io pk is the function to be called when interactively importing a key-pair and returning the public key and its hash.

Some signer implementations improve long-term security by requiring human/manual validation while importing keys, the ~io argument can be used to prompt the user in such case.

val public_key : pk_uri -> (Tezos_crypto.Signature.Public_key.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

public_key pk is the Ed25519 version of pk.

val public_key_hash : pk_uri -> (Tezos_crypto.Signature.Public_key_hash.t * Tezos_crypto.Signature.Public_key.t option, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

public_key_hash pk is the hash of pk. As some signers will query the full public key to obtain the hash, it can be optionally returned to reduce the amount of queries.

val sign : ?watermark:Tezos_crypto.Signature.watermark -> sk_uri -> Stdlib.Bytes.t -> (Tezos_crypto.Signature.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

sign ?watermark sk data is signature obtained by signing data with sk.

val deterministic_nonce : sk_uri -> Stdlib.Bytes.t -> (Stdlib.Bytes.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

deterministic_nonce sk data is a nonce obtained deterministically from data and sk.

val deterministic_nonce_hash : sk_uri -> Stdlib.Bytes.t -> (Stdlib.Bytes.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

deterministic_nonce_hash sk data is a nonce hash obtained deterministically from data and sk.

val supports_deterministic_nonces : sk_uri -> (bool, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

supports_deterministic_nonces indicates whether the deterministic_nonce functionality is supported.