Module Tezos_client_base__Client_context

This module contains several class types used to provide I/O services for the client, the wallet, etc.

This interface is independent of the backend. This is intended for the client library code, so it can be independent of the backend. The client, and each other client-like systems (wallet apps, in-browser widgets, etc.), are responsible for providing a backend-specific implementation.

To find examples on how these methods are called, you can have a look at

src/bin_client/client_protocols_commands.ml

src/lib_client_base/client_aliases.ml

src/lib_client_base_unix/client_context_unix.ml